==> Codes/01_train_test_split.ipynb <==
import pandas as pd
import os
from sklearn.model_selection import train_test_split
import numpy as np

os.chdir('../Input_Data')

df = pd.read_csv('train.csv',header=0)

y = np.array(df['Survived'])

x = np.array(df.drop(['Survived'],axis=1))

x_train, x_test, y_train, y_test = train_test_split(x, y, test_size=0.2, random_state=42)

df_train = pd.DataFrame(x_train,columns=df.drop(['Survived'],axis=1).columns)

df_test = pd.DataFrame(x_test,columns=df.drop(['Survived'],axis=1).columns)

df_train['Survived'] = y_train

df_test['Survived'] = y_test

os.chdir('../output_data')

df_train.to_csv('train.csv',index=False)

df_test.to_csv('validation.csv',index=False)
