==> Codes/04_variable_reduction.ipynb <==
import pandas as pd
import numpy as np
import pickle

from sklearn.linear_model import LinearRegression
from sklearn.metrics import r2_score

from scipy.stats import skew,kurtosis

import os

os.chdir('../pickles')

numerical=pickle.load(open('numerical.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))

os.chdir('../output_data')

df = pd.read_csv('train_v03.csv',header=0)

df.set_index('PassengerId',inplace=True)

def edd(data):
    df_desc = data.describe().transpose()
    df_desc['Var'] = df_desc.index
    df_desc.reset_index(inplace=True)
    df_desc.drop('count',axis=1,inplace=True)
    df_desc['skewness'] = df_desc['Var'].apply(lambda x: skew(np.array(data.loc[data[x].notnull(),x])))
    df_desc['kurtosis'] = df_desc['Var'].apply(lambda x: kurtosis(np.array(data.loc[data[x].notnull(),x]),fisher=False))
    df_desc['99%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.99))
    df_desc['95%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.95))
    df_desc['90%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.90))
    df_desc['10%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.1))
    df_desc['5%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.05))
    df_desc['1%'] = df_desc['Var'].apply(lambda x: data[x].quantile(.01))
    df_desc['mean-3sigma'] = df_desc['mean'] - 3*df_desc['std']
    df_desc['mean+3sigma'] = df_desc['mean'] + 3*df_desc['std']
    df_desc['mean-2sigma'] = df_desc['mean'] - 2*df_desc['std']
    df_desc['mean+2sigma'] = df_desc['mean'] + 2*df_desc['std']
    df_desc['type']='numeric'
    
    def is_category(col):
        return 'float' not in str(data[col].dtype) and 'int' not in str(data[col].dtype)
    
    categorical = [col for col in data.columns if is_category(col)]
    df_categorical = pd.DataFrame()
    df_categorical['Var']=np.array(categorical)
    
    df_categorical['type']='categorical'
    for col in [c for c in df_desc.columns if c not in ['Var','type']]:
        df_categorical[col]=np.nan
    for col in categorical:
        df_var = data[col].value_counts()
        df_cat = pd.DataFrame()
        df_cat['count']=df_var
        df_cat['categories']=df_var.index
        df_cat.reset_index(inplace=True)
        df_cat.sort_values(by='count',ascending=False,inplace=True)
        df_cat.set_index('categories',inplace=True)
        index_list = df_cat.index.tolist()
        for i,c in enumerate(['mean','min','1%','5%','10%','25%']):
            try:
                df_categorical.loc[df_categorical['Var']==col,c] = index_list[i]
            except:
                break
        for i,c in enumerate(['50%','75%','90%','95%','99%','max']):
            try:
                df_categorical.loc[df_categorical['Var']==col,c] = index_list[-(i+1)]
            except:
                break
        del df_var
        del df_cat
        del index_list
    df_categorical = df_categorical[df_desc.columns]
    edd = pd.concat([df_desc,df_categorical])
    del df_desc
    del df_categorical
    edd['count'] = edd['Var'].apply(lambda x: data[data[x].notnull()].shape[0])
    edd['nmiss'] = data.shape[0]-edd['count']
    edd['missing_rate'] = np.array(edd['nmiss']).astype('float')/data.shape[0] * 100
    edd['unique'] = edd['Var'].apply(lambda x: len(data[x].value_counts().index.tolist()))
    orig_cols = ['mean','min','1%','5%','10%','25%','50%','75%','90%','95%','99%','max']
    new_cols = ['mean_or_top1','min_or_top2','p1_or_top3','p5_or_top4','p10_or_top5','p25_or_top6',
                'p50_or_bottom6','p75_or_bottom5','p90_or_bottom4','p95_or_bottom3','p99_or_bottom2','max_or_bottom1']
    
    convert_dict = {}
    for i in range(len(orig_cols)):
        convert_dict[orig_cols[i]]=new_cols[i]
    edd.rename(columns=convert_dict,inplace=True)
    edd = edd[['Var','type','count','nmiss','missing_rate','unique','std','skewness','kurtosis','mean-3sigma',
               'mean-2sigma','mean_or_top1','min_or_top2','p1_or_top3','p5_or_top4','p10_or_top5','p25_or_top6',
               'p50_or_bottom6','p75_or_bottom5','p90_or_bottom4','p95_or_bottom3','p99_or_bottom2','max_or_bottom1'
              ,'mean+2sigma','mean+3sigma']]
    return edd

edd =edd(df)

from scipy.stats import chisquare
from statsmodels.stats.weightstats import ztest

def p_value_stat(col,target='Survived'):
    if col==target:
        return None
    try:
        if edd.loc[edd['Var']==col,'type'].values[0]=='categorical':
            obs_freq = []
            for cat in df[col].value_counts().index.tolist():
                obs_class=[]
                for i in df[target].value_counts().index.tolist():
                    obs = df[(df[col]==str(cat)) & (df[target]==i)].shape[0]
                    obs_class.append(obs)
                obs_freq.append(obs_class)
            obs_freq=np.array(obs_freq)
            return chisquare(obs_freq,axis=0)[1][0]
        else:
            return ztest(np.array(df.loc[(df[col].notnull()) & (df[target]==1),col]),
                        np.array(df.loc[(df[col].notnull()) & (df[target]==0),col]))[1]
    except Exception as e:
        print(e,col)
        return np.nan

edd['Anova/Chisquare pvalue'] = edd['Var'].apply(lambda x: p_value_stat(x))

os.chdir('../Statistics')

edd.to_csv('edd_after_treatment.csv',index=False)

cols_to_drop =list(edd.loc[edd['unique']<2,'Var'])

df.drop(cols_to_drop,axis=1,inplace=True)

def inter_correlation_clusters(data,cutoff=.7):
    correlations=data.corr()
    graph={}
    columns=data.columns
    for i in range(len(columns)):
        graph[i]=[]
        for j in range(len(columns)):
            if i!=j and np.abs(correlations.iloc[i,j])>cutoff:
                graph[i].append(j)
    
    tree_set={}
    component = 0
    visited = [0 for i in range(len(columns))]
    def dfs(i):
        visited[i]=1
        try:
            tree_set[component].append(i)
        except KeyError:
            tree_set[component] = [i]
            
        for j in graph[i]:
            if visited[j]==0:
                dfs(j)
                
    tree_cluster={}
    for key in list(tree_set.keys()):
        tree_cluster[key] = [columns[i] for i in tree_set[key]]
        
    return tree_cluster

def varclus(data):
    columns = []
    correlations=data.corr()
    clusters= inter_correlation_clusters(data)
    print(clusters)
    
    cols = list(data.columns)
    
    def distance(c1,c2):
        return np.max([[np.abs(correlations.loc[i,j]) for i in clusters[c1]] for j in clusters[c2]])
    
    def next_closest(c):
        minima=0
        point=c
        for c1 in [i for i in list(clusters.keys()) if i!=c]:
            dist = distance(c,c1)
            if dist>minima:
                minima=dist
                point=c1
        return point
    
    def get_squared_ratio(col,own_cluster,next_cluster):
        y = np.array(data[col])
        x = np.array(data[own_cluster].drop(col,axis=1))
        model=LinearRegression()
        model.fit(x,y)
        y_pred = list(model.predict(x))
        del x
        del model
        r2_own = r2_score(y,y_pred)
        del y_pred
        x = np.array(data[next_cluster])
        model=LinearRegression()
        model.fit(x,y)
        y_pred = list(model.predict(x))
        del x
        del model
        r2_next = r2_score(y,y_pred)
        del y
        del y_pred
        
        return float(1-r2_own)/(1-r2_next)
    
    for c1 in list(clusters.keys()):
        if len(clusters[c1])>1:
            own_cluster = clusters[c1]
            next_cluster = clusters[next_closest(c1)]
            ration=np.inf
            for col in clusters[c1]:
                col_ratio = get_squared_ratio(col,own_cluster,next_cluster)
                if col_ratio<ratio:
                    ratio=col_ratio
                    clust_col=col
            columns.append(clust_col)
        else:
            columns.append(clusters[c1][0])
            
    return columns

columns = varclus(df.drop(['Survived'],axis=1))

for col in df.columns:
    if True in np.isinf(np.array(df[col])):
        print col

# Variance Inflation Factor

df.shape

from patsy import dmatrices
import statsmodels
from statsmodels.stats.outliers_influence import variance_inflation_factor

def variance_inflation(data):
    vif = pd.DataFrame()
    vif['features'] = data.columns
    vif['vif factor'] = [variance_inflation_factor(data.values, i) for i in range(data.shape[1])]
    vif.sort_values(by='vif factor',ascending=False,inplace=True)
    vif.reset_index(inplace=True)
    vif.drop(['index'],axis=1,inplace=True)
    print(vif)
    return tuple(vif.loc[0,:].values)

vif_drop_cols=[]

def vif_reduction(data,limit=2.5):
    vif = variance_inflation(data)
    if vif[1]<=limit:
        return
    else:
        data.drop(vif[0],axis=1,inplace=True)
        vif_drop_cols.append(vif[0])
        print(vif[0]+' dropped')
        del vif
        vif_reduction(data,limit=limit)

vif_reduction(df.drop(['Survived'],axis=1))

len(vif_drop_cols)

os.chdir('../Statistics')

edd = pd.read_csv('edd_v03.csv',header=0)

edd.loc[edd['Var'].isin(vif_drop_cols),'Status']='drop'

edd['Reason']=''
edd.loc[edd['Var'].isin(vif_drop_cols),'Reason']='VIF'

edd.to_csv('edd_v04.csv',index=False)

df.drop(vif_drop_cols,axis=1,inplace=True)

os.chdir('../output_data')

df.to_csv('train_v04.csv')

# BackWard Selection

import statsmodels.discrete.discrete_model as sm
from sklearn.preprocessing import StandardScaler

flag=0
cols_dropped=['Survived']
while flag==0:
    model = sm.Logit(endog=np.array(df['Survived']),exog=np.array(df.drop(cols_dropped,axis=1)))
    results = model.fit()
    pvalues=list(results.pvalues)
    drop_index = pvalues.index(max(pvalues))
    col_drop = df.drop(cols_dropped,axis=1).columns[drop_index]
    print(col_drop+'-'+str(pvalues[drop_index]))
    if pvalues[drop_index]>.05:
        cols_dropped.append(col_drop)
    else:
        flag=1

cols_dropped.remove('Survived')

len(cols_dropped)

df.shape

df.drop(cols_dropped,axis=1,inplace=True)

df.columns

os.chdir('../output_data')

df.to_csv('train_v05.csv')

os.chdir('../pickles')

pickle.dump(vif_drop_cols,open('vif_drop_cols.pickle','wb'))

pickle.dump(cols_dropped,open('backward_drop_cols.pickle','wb'))

df.shape

df['Survived']

os.chdir('../Statistics')

edd = pd.read_csv('edd_v04.csv',header=0)

edd.loc[edd['Var'].isin(cols_dropped),'Status']='drop'

edd.loc[edd['Var'].isin(cols_dropped),'Reason']='Backward Selection'

edd.to_csv('edd_v05.csv',index=False)

os.chdir('../pickles')

numericals_final = list(set(set(df.columns) & set(numerical)))

len(numericals_final)

numericals_final

dummies_final = [col for col in df.columns if 'dum' in col]

len(dummies_final)

pickle.dump(numericals_final,open('numerical_final.pickle','wb'))
pickle.dump(dummies_final,open('dummies_final.pickle','wb'))

# PCA

from sklearn.decomposition import PCA

from sklearn.preprocessing import StandardScaler

scale = StandardScaler()

scale.fit(np.array(df.drop(['Survived'],axis=1)))

pickle.dump(scale,open('scale.pickle','wb'))

x = scale.transform(df.drop(['Survived'],axis=1))

df.shape

pca = PCA(n_components=9)

pca.fit(x)

variance_cumulative = np.cumsum(pca.explained_variance_ratio_)

for i in range(61):
    if variance_cumulative[i]>=.9:
        n_components=i+1
        break

n_components

pca = PCA(n_components=n_components)

pca.fit(x)

pickle.dump(pca,open('pca.pickle','wb'))

df_pca_weights = pd.DataFrame(pca.components_,columns=df.drop(['Survived'],axis=1).columns,index=['pc'+str(i)
                                                                                                  for i in range(1,n_components+1)])

df_pca_magnitude_weights = pd.DataFrame(map(np.square,pca.components_),columns=df.drop(['Survived'],axis=1).columns,index=['pc'+str(i)
                                                                                                  for i in range(1,n_components+1)])

df_pca_weights

os.chdir('../PCA')

df_pca_weights.to_csv('pca_weights.csv')

df_pca_magnitude_weights.to_csv('pca_magnitude_weights.csv')

df_pca = pd.DataFrame(pca.transform(x),columns=['pc'+str(i) for i in range(1,n_components+1)])

df_pca['Survived']=df['Survived']

os.chdir('../output_data')

df_pca.index = df.index

df_pca['Survived'].isnull().any()

df_pca.to_csv('train_pca.csv')
