==> Codes/07_predictions.ipynb <==
import pandas as pd
import numpy as np
import pickle
import os

from scipy.special import cbrt
import re

os.chdir('../pickles')

numerical = pickle.load(open('numerical_final.pickle','rb'))
categorical = pickle.load(open('categorical.pickle','rb'))
conversion_dict = pickle.load(open('conversion_dict.pickle','rb'))
scale = pickle.load(open('scale.pickle','rb'))
imputation_cols = pickle.load(open('imputation_cols.pickle','rb'))
dummy_cols = pickle.load(open('dummies_final.pickle','rb'))
pca = pickle.load(open('pca.pickle','rb'))

model_cols = pickle.load(open('model_columns.pickle'))

os.chdir('../Input_Data')

df = pd.read_csv('test.csv',header=0)

df['Cabin']=df['Cabin'].astype(str).apply(lambda x: ''.join([s for s in x if str.isalpha(s)][0]))
df['Ticket']=df['Ticket'].apply(lambda x: ''.join([s for s in x if str.isalpha(s)]))
def get_title(name):
    title_search = re.search(' ([A-Za-z]+)\.', name)
    # If the title exists, extract and return it.
    if title_search:
        return title_search.group(1)
    return ""
df['Pclass'] = df['Pclass'].astype(str).apply(lambda x: x+' class')
df['Title'] = df['Name'].apply(get_title) 
df.drop(['Name'],axis=1,inplace=True)
df = df[numerical+categorical+['PassengerId']]
df.set_index('PassengerId',inplace=True)

values={}
for col in categorical:
    values[col]='Unknown'
df.fillna(values,inplace=True)

for col in categorical:
    if len(conversion_dict[col])<=2:
        for i in range(2):
            try:
                category = conversion_dict[col][i]
                df[col+'_dum_'+str(category)] = 0
                df.loc[df[col]==category,col+'_dum_'+str(category)]=1
            except:
                pass
    else:
        total_categories = len(conversion_dict[col])
        dummies = len(str(int(bin(total_categories)[2:],10)))
        bin_conv=[]
        for i in range(total_categories):
            bin_conv.append(conversion_dict[col][i][1])
        for j in range(dummies):
            df[col+'_dum_'+str(j)]=0
            for i,cat in enumerate([conv[0] for conv in conversion_dict[col]]):
                df.loc[df[col]==cat,col+'_dum_'+str(j)]=bin_conv[i]%10
                bin_conv[i]=bin_conv[i]//10
    df.drop(col,axis=1,inplace=True)
    print(col+' done')   

transform_dict = {'log':lambda x: np.log(x),'sqr':lambda x: x**2,'sqrt':lambda x: np.sqrt(x),'exp':lambda x:np.exp(x),
                 'cube':lambda x: x**3,'cuberoot': lambda x: cbrt(x)}

os.chdir('../Statistics')

edd = pd.read_csv('edd_v05.csv',header=0)

def transform(x,function):
    if x is not None:
        return function(x)
    else:
        return np.nan

for col in numerical:
    if edd[(edd['Var']==col)&(edd['Status'].notnull())].shape[0]!=0:
        function = transform_dict[edd.loc[edd['Var']==col,'Status'].values[0]]
        df[col] = df[col].apply(lambda x: transform(x,function))

os.chdir('../Imputation_models')

df[imputation_cols]

for col in ['Fare']:
    model = pickle.load(open(col+'_impute.pickle','rb'))
    if df[col].isnull().any():
        indices = df.loc[df[col].isnull()].index.tolist()
        df.loc[indices,col]=np.array(model.predict(np.array(df.loc[indices,imputation_cols])))
        del indices
    del model
    print(col+' imputed')

df = df[numerical+dummy_cols]

# Input Arrays

x = np.array(df[model_cols])

x_pca = pca.transform(scale.transform(x))

# Automated Model Prediction

def submission(model_name,x,df_eval,model_dict):
    if 'predictive' in model_name:
        x = pca.transform(scale.transform(x))
    def predict(prob,cutoff):
        if prob >= cutoff:
            return 1
        else:
            return 0
    if 'ensemble' in model_name:
        model_list = model_name.split('(')[1].split(')')[0].split(',')
        models = [(model_nm,model_dict[model_nm]) for model_nm in model_list]
        probabilities = sum([model[1].predict_proba(x)[:,1]*df_eval.loc[df_eval['model_name']==model[0],'roc_train'].values[0]
                             for model in models])/sum([df_eval.loc[df_eval['model_name']==model[0],'roc_train'].values[0] for model 
                                                        in models])  
        cutoff = df_eval.loc[df_eval['model_name']==model_name,'cutoff'].values[0]
        return [predict(prob,cutoff) for prob in probabilities]
    
    else:
        model = model_dict[model_name]
        probabilities = model.predict_proba(x)[:,1]
        cutoff = df_eval.loc[df_eval['model_name']==model_name,'cutoff'].values[0]
        return [predict(prob,cutoff) for prob in probabilities]

os.chdir('../Models/Inferential_models')

model_dict = {'LR_inferential':pickle.load(open('LR.pickle','rb')),
                      'RF_inferential':pickle.load(open('RF.pickle','rb')),
                    'GBM_inferential':pickle.load(open('GBM.pickle','rb'))}

os.chdir('../pca_models')

model_dict['LR_predictive']=pickle.load(open('LR.pickle','rb'))
model_dict['RF_predictive']=pickle.load(open('RF.pickle','rb'))
model_dict['GBM_predictive']=pickle.load(open('GBM.pickle','rb'))

os.chdir('../../Statistics')

df_eval = pd.read_csv('model_evaluation.csv',header=0)

os.chdir('../Submissions')

for i,model_name in enumerate(list(df_eval['model_name'])):
    df_submission = pd.DataFrame()
    model_name = df_eval.loc[i,'model_name']
    df_submission['Survived'] = submission(model_name,x,df_eval,model_dict)
    df_submission.index = df.index
    df_submission.to_csv('submission_'+str(i+1)+'.csv')
